♻️ (Documentation component) - Move documentation component files in app folder
